+++ src/components/AnswerPageQuestionItem/style.js
@@ -65,31 +65,9 @@ const QuestionContainer = styled.div`
     background: var(--Grayscale-10);
   }
 
-  .write-date {
-    color: var(--Grayscale-40);
-    font-feature-settings:
-      'clig' off,
-      'liga' off;
-    font-size: 1.4rem;
-    font-weight: 500;
-    line-height: 18px;
-
-    margin-bottom: 4px;
-  }
-
   .card-answer-section {
     display: flex;
     gap: 12px;
-  }
-
-  .card-title {
-    color: var(--Grayscale-60);
-    font-feature-settings:
-      'clig' off,
-      'liga' off;
-    font-family: Actor;
-    font-size: 1.8rem;
-    line-height: 24px;
   }
 
   .card-profile-wrap {
