+++ src/pages/QuestionPage/index.js
@@ -1,21 +1,15 @@
 import { useEffect, useState, useCallback } from 'react';
-import { Link, useParams } from 'react-router-dom';
+import { useParams } from 'react-router-dom';
 import useInfiniteScroll from '../../hooks/useInfiniteScroll';
 import getElapsedTime from '../../utils/getElapsedTime';
 import ReactionButtonBox from '../../components/ReactionButtonBox';
 import ModalQuestion from '../../components/ModalQuestion';
-import ShareButtonBar from '../../components/ShareButtonBar';
 import QuestionPageContainer from './style';
+import QuestionHeader from '../../components/QuestionHeader';
+import getUserData from '../../api/getUserData';
+import QuestionCardHeader from '../../components/QuestionCardHeader';
 
 const API_BASE_URL = 'https://openmind-api.vercel.app/3-5';
-
-const getUser = async userId => {
-  const response = await fetch(`${API_BASE_URL}/subjects/${userId}/`);
-  if (!response.ok) {
-    throw new Error('유저 데이터를 불러오는데 실패했습니다');
-  }
-  return response.json();
-};
 
 const QuestionItem = ({ user, question }) => {
   return (
@@ -28,15 +22,7 @@ const QuestionItem = ({ user, question }) => {
       )}
       {!question.answer && <div className="answer none">미답변</div>}
 
-      <div className="question-box">
-        <div className="question-title-box">
-          <span className="question-title">질문 · </span>
-          <span className="question-time-line">
-            {getElapsedTime(question.createdAt)}
-          </span>
-        </div>
-        <span className="question-content">{question.content}</span>
-      </div>
+      <QuestionCardHeader question={question} />
 
       {question.answer ? (
         <div className="answer-box">
@@ -63,31 +49,6 @@ const QuestionItem = ({ user, question }) => {
 
       <ReactionButtonBox question={question} />
     </section>
-  );
-};
-
-const QuestionHeader = ({ user }) => {
-  return (
-    <header className="questions-page-header">
-      <section className="header-section">
-        <Link to="/">
-          <img
-            src="/assets/images/logo.png"
-            className="logo"
-            alt="OpenMind 서비스 로고"
-          />
-        </Link>
-        <section className="profile">
-          <img
-            src={user.imageSource}
-            className="header-profile-image"
-            alt="프로필 이미지"
-          />
-          <span className="profile-name">{user.name}</span>
-          <ShareButtonBar />
-        </section>
-      </section>
-    </header>
   );
 };
 
@@ -122,7 +83,7 @@ const QuestionPage = () => {
   const elementRef = useInfiniteScroll(getUserQuestions);
 
   const fetchData = async () => {
-    const responseUser = await getUser(id);
+    const responseUser = await getUserData(id);
     setUser(responseUser);
     setQuestionCount(responseUser.questionCount);
   };
