+++ src/pages/QuestionPage/style.js
@@ -7,80 +7,6 @@ const QuestionPageContainer = styled.div`
     cursor: pointer;
   }
 
-  .questions-page-header {
-    position: relative;
-    height: 234px;
-    background-color: var(--Grayscale-10);
-  }
-
-  .header-section {
-    position: absolute;
-    bottom: 0;
-    background-image: url(/assets/images/logoWithoutBg.png);
-    background-repeat: no-repeat;
-    background-position: center top -10%;
-    width: 100%;
-    height: 184px;
-    display: flex;
-    align-items: center;
-    flex-direction: column;
-    gap: 12px;
-
-    .logo {
-      width: 170px;
-      height: 67px;
-    }
-
-    .profile {
-      display: flex;
-      flex-direction: column;
-      align-items: center;
-      gap: 12px;
-    }
-
-    .header-profile-image {
-      width: 136px;
-      height: 136px;
-      border-radius: 9999px;
-    }
-
-    .profile-name {
-      font-size: 3.2rem;
-    }
-
-    .profile-share-icons {
-      display: flex;
-      gap: 12px;
-    }
-
-    .profile-share-link-box {
-      width: 40px;
-      height: 40px;
-      border-radius: 9999px;
-      cursor: pointer;
-      display: flex;
-      justify-content: center;
-      align-items: center;
-    }
-
-    .profile-share-link-logo {
-      width: 18px;
-      height: 18px;
-    }
-
-    .profile-share-link-box.brown {
-      background-color: var(--Brown-40);
-    }
-
-    .profile-share-link-box.yellow {
-      background-color: var(--Yellow-50);
-    }
-
-    .profile-share-link-box.blue {
-      background-color: var(--Blue-50);
-    }
-  }
-
   .questions-page-main {
     display: flex;
     justify-content: center;
@@ -191,24 +117,6 @@ const QuestionPageContainer = styled.div`
     .answer.rejected {
       border-color: var(--Red-50);
       color: var(--Red-50);
-    }
-
-    .question-box {
-      display: flex;
-      flex-direction: column;
-      gap: 4px;
-    }
-
-    .question-title-box {
-      font-size: 1.4rem;
-      color: var(--Grayscale-40);
-      margin-bottom: 4px;
-    }
-
-    .question-content {
-      font-size: 1.8rem;
-      font-family: Actor;
-      white-space: pre-line;
     }
 
     .answer-box {
